+++ dubbo-test/dubbo-dependencies-all/pom.xml
@@ -186,6 +186,10 @@
         </dependency>
         <dependency>
             <groupId>org.apache.dubbo</groupId>
+            <artifactId>dubbo-spring-security</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-reactive</artifactId>
         </dependency>
